+++ src/Content/PortalNode/PortalNodeDefinition.php
@@ -14,6 +14,9 @@ use Shopware\Core\Framework\DataAbstractionLayer\Field\OneToManyAssociationField
 use Shopware\Core\Framework\DataAbstractionLayer\Field\StringField;
 use Shopware\Core\Framework\DataAbstractionLayer\FieldCollection;
 
+/**
+ * @deprecated DAL usage is discouraged. Use portal node specific actions instead
+ */
 class PortalNodeDefinition extends EntityDefinition
 {
     public const ENTITY_NAME = 'heptaconnect_portal_node';